Create file for place id request
